+++ deepinv/optim/__init__.py
@@ -1,5 +1,5 @@
 from .data_fidelity import DataFidelity, L2, L1, IndicatorL2, PoissonLikelihood
 from .optimizers import BaseOptim, optim_builder
 from .fixed_point import FixedPoint
-from .prior import Prior, ScorePrior, Tikhonov, PnP, RED, L1Prior
+from .prior import Prior, ScorePrior, Tikhonov, PnP, RED, L1Prior, L21Prior, TVPrior
 from .optim_iterators.optim_iterator import OptimIterator